New spout example
