+++ examples/spout-marker/producer/Dockerfile
@@ -1,0 +1,4 @@
+FROM python:3
+ADD requirements.txt /app/requirements.txt
+ADD main.py /app/main.py
+RUN pip3 install -r /app/requirements.txt
