+++ examples/spout-marker/producer/requirements.txt
@@ -1,0 +1,1 @@
+python-pachyderm>=4.0.0
