+++ examples/spout/producer/Makefile
@@ -1,6 +1,6 @@
 .PHONY: docker-build
 
 docker-build:
-	docker build -t pachyderm_spout_producer --no-cache .
-	docker tag pachyderm_spout_producer ysimonson/pachyderm_spout_producer
-	docker push ysimonson/pachyderm_spout_producer
+	docker build -t pachyderm_example_spout_producer --no-cache .
+	docker tag pachyderm_example_spout_producer ysimonson/pachyderm_example_spout_producer
+	docker push ysimonson/pachyderm_example_spout_producer
